+++ tests/CLI/modules/sshkey_tests.py
@@ -6,6 +6,7 @@
 """
 import json
 import os.path
+import sys
 import tempfile
 
 import mock
@@ -102,6 +103,8 @@ class SshKeyTests(testing.TestCase):
                          {'id': 1234, 'label': 'label', 'notes': 'notes'})
 
     def test_print_key_file(self):
+        if(sys.platform.startswith("win")):
+            self.skipTest("Test doesn't work in Windows")
         with tempfile.NamedTemporaryFile() as sshkey_file:
             service = self.client['Security_Ssh_Key']
             mock_key = service.getObject()['key']