+++ tests/managers/block_tests.py
@@ -5,6 +5,7 @@
     :license: MIT, see LICENSE for more details.
 """
 
+import copy
 import SoftLayer
 from SoftLayer import exceptions
 from SoftLayer import fixtures
@@ -17,6 +18,21 @@ class BlockTests(testing.TestCase):
 
     def test_cancel_block_volume_immediately(self):
         self.block.cancel_block_volume(123, immediate=True)
+
+        self.assert_called_with(
+            'SoftLayer_Billing_Item',
+            'cancelItem',
+            args=(True, True, 'No longer needed'),
+            identifier=449,
+        )
+
+    def test_cancel_block_volume_immediately_hourly_billing(self):
+        mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
+        mock.return_value = {
+            'billingItem': {'hourlyFlag': True, 'id': 449},
+        }
+
+        self.block.cancel_block_volume(123, immediate=False)
 
         self.assert_called_with(
             'SoftLayer_Billing_Item',
@@ -181,6 +197,48 @@ class BlockTests(testing.TestCase):
             identifier=123,
         )
 
+    def test_cancel_snapshot_hourly_billing_immediate_true(self):
+        mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
+        mock.return_value = {
+            'billingItem': {
+                'activeChildren': [
+                    {'categoryCode': 'storage_snapshot_space', 'id': 417}
+                ],
+                'hourlyFlag': True,
+                'id': 449
+            },
+        }
+
+        self.block.cancel_snapshot_space(1234, immediate=True)
+
+        self.assert_called_with(
+            'SoftLayer_Billing_Item',
+            'cancelItem',
+            args=(True, True, 'No longer needed'),
+            identifier=417,
+        )
+
+    def test_cancel_snapshot_hourly_billing_immediate_false(self):
+        mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
+        mock.return_value = {
+            'billingItem': {
+                'activeChildren': [
+                    {'categoryCode': 'storage_snapshot_space', 'id': 417}
+                ],
+                'hourlyFlag': True,
+                'id': 449
+            },
+        }
+
+        self.block.cancel_snapshot_space(1234, immediate=False)
+
+        self.assert_called_with(
+            'SoftLayer_Billing_Item',
+            'cancelItem',
+            args=(True, True, 'No longer needed'),
+            identifier=417,
+        )
+
     def test_cancel_snapshot_exception_no_billing_item_active_children(self):
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = {
@@ -279,8 +337,7 @@ class BlockTests(testing.TestCase):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_storage_type_keyname = mock_volume['storageType']['keyName']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -313,11 +370,10 @@ class BlockTests(testing.TestCase):
                 'quantity': 1,
                 'location': 449494,
                 'iops': 2000,
+                'useHourlyPricing': False,
                 'osFormatType': {'keyName': 'LINUX'}
             },)
         )
-
-        mock_volume['storageType']['keyName'] = prev_storage_type_keyname
 
     def test_order_block_volume_endurance(self):
         mock = self.set_mock('SoftLayer_Location_Datacenter', 'getDatacenters')
@@ -357,6 +413,7 @@ class BlockTests(testing.TestCase):
                 'volumeSize': 1000,
                 'quantity': 1,
                 'location': 449494,
+                'useHourlyPricing': False,
                 'osFormatType': {'keyName': 'LINUX'}
             },)
         )
@@ -439,8 +496,7 @@ class BlockTests(testing.TestCase):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_storage_type_keyname = mock_volume['storageType']['keyName']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'ENDURANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -461,11 +517,10 @@ class BlockTests(testing.TestCase):
                 ],
                 'quantity': 1,
                 'location': 449500,
-                'volumeId': 102
+                'volumeId': 102,
+                'useHourlyPricing': False
             },)
         )
-
-        mock_volume['storageType']['keyName'] = prev_storage_type_keyname
 
     def test_order_block_snapshot_space(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
@@ -491,13 +546,13 @@ class BlockTests(testing.TestCase):
                 ],
                 'quantity': 1,
                 'location': 449500,
-                'volumeId': 102
+                'volumeId': 102,
+                'useHourlyPricing': False
             },)
         )
 
     def test_order_block_replicant_os_type_not_found(self):
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_os_type = mock_volume['osType']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         del mock_volume['osType']
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -514,8 +569,6 @@ class BlockTests(testing.TestCase):
             str(exception)
         )
 
-        mock_volume['osType'] = prev_os_type
-
     def test_order_block_replicant_performance_os_type_given(self):
         mock = self.set_mock('SoftLayer_Location_Datacenter', 'getDatacenters')
         mock.return_value = [{'id': 449494, 'name': 'dal09'}]
@@ -523,8 +576,7 @@ class BlockTests(testing.TestCase):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_storage_type_keyname = mock_volume['storageType']['keyName']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -559,11 +611,10 @@ class BlockTests(testing.TestCase):
                 'iops': 1000,
                 'originVolumeId': 102,
                 'originVolumeScheduleId': 978,
+                'useHourlyPricing': False,
                 'osFormatType': {'keyName': 'XEN'}
             },)
         )
-
-        mock_volume['storageType']['keyName'] = prev_storage_type_keyname
 
     def test_order_block_replicant_endurance(self):
         mock = self.set_mock('SoftLayer_Location_Datacenter', 'getDatacenters')
@@ -600,6 +651,7 @@ class BlockTests(testing.TestCase):
                 'location': 449494,
                 'originVolumeId': 102,
                 'originVolumeScheduleId': 978,
+                'useHourlyPricing': False,
                 'osFormatType': {'keyName': 'LINUX'}
             },)
         )
@@ -608,8 +660,7 @@ class BlockTests(testing.TestCase):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_os_type = mock_volume['osType']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         del mock_volume['osType']
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -623,14 +674,11 @@ class BlockTests(testing.TestCase):
         self.assertEqual(str(exception),
                          "Cannot find origin volume's os-type")
 
-        mock_volume['osType'] = prev_os_type
-
     def test_order_block_duplicate_performance_no_duplicate_snapshot(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_storage_type_keyname = mock_volume['storageType']['keyName']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -659,17 +707,15 @@ class BlockTests(testing.TestCase):
                 'location': 449500,
                 'duplicateOriginVolumeId': 102,
                 'osFormatType': {'keyName': 'LINUX'},
-                'iops': 1000
+                'iops': 1000,
+                'useHourlyPricing': False
             },))
-
-        mock_volume['storageType']['keyName'] = prev_storage_type_keyname
 
     def test_order_block_duplicate_performance(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
         mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
-        prev_storage_type_keyname = mock_volume['storageType']['keyName']
+        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -705,10 +751,9 @@ class BlockTests(testing.TestCase):
                 'duplicateOriginVolumeId': 102,
                 'osFormatType': {'keyName': 'LINUX'},
                 'duplicateOriginSnapshotId': 470,
-                'iops': 2000
+                'iops': 2000,
+                'useHourlyPricing': False
             },))
-
-        mock_volume['storageType']['keyName'] = prev_storage_type_keyname
 
     def test_order_block_duplicate_endurance_no_duplicate_snapshot(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
@@ -741,7 +786,8 @@ class BlockTests(testing.TestCase):
                 'quantity': 1,
                 'location': 449500,
                 'duplicateOriginVolumeId': 102,
-                'osFormatType': {'keyName': 'LINUX'}
+                'osFormatType': {'keyName': 'LINUX'},
+                'useHourlyPricing': False
             },))
 
     def test_order_block_duplicate_endurance(self):
@@ -782,7 +828,8 @@ class BlockTests(testing.TestCase):
                 'location': 449500,
                 'duplicateOriginVolumeId': 102,
                 'osFormatType': {'keyName': 'LINUX'},
-                'duplicateOriginSnapshotId': 470
+                'duplicateOriginSnapshotId': 470,
+                'useHourlyPricing': False
             },))
 
     def test_setCredentialPassword(self):